Improved state structure
